feat: add request for user
