+++ SocialNetworkApi/Mappers/Request/Post/EditPostRequest.cs
@@ -1,4 +1,4 @@
-using SocialNetwork.Domain;
+using SocialNetwork.Entities;
 
 namespace SocialNetwork.Mappers.Requests;
 public record EditPostRequest(string Content, string ImageUrl)
